+++ apps/site-app/src/containers/Widgets/Balance/BalanceWidget.stories.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import { Meta } from '@storybook/react';
+import { BalanceWidget as Component } from '.';
+import styles from '../../containers/App/styles.module.less';
+import { withAccounts } from '@thecointech/storybookutils';
+
+export default {
+  title: 'App/Widgets/Balance',
+  component: Component,
+  decorators: [withAccounts()]
+} as Meta;
+
+export const Balance = () => (
+  <div id={styles.app}>
+    <Component />
+  </div>
+);
